Move tests to the correct location
