+++ tests/multimodal/test_inputs.py
@@ -4,10 +4,7 @@
 import pytest
 import torch
 
-from vllm.config import ModelConfig
-from vllm.inputs.preprocess import InputPreprocessor
 from vllm.multimodal.inputs import MultiModalKwargs, NestedTensors
-from vllm.transformers_utils.tokenizer import init_tokenizer_from_configs
 
 pytestmark = pytest.mark.cpu_test
 
@@ -102,50 +99,3 @@ def test_multimodal_input_batch_mixed_stacking_depths():
 
     result = MultiModalKwargs.batch([{"image": [a]}, {"image": [b, c]}])
     assert_multimodal_inputs_equal(result, {"image": [a.unsqueeze(0), [b, c]]})
-
-
-@pytest.mark.parametrize("model_id", [
-    "facebook/opt-125m",
-])
-@pytest.mark.parametrize("prompt", [
-    {
-        "prompt": "",
-        "multi_modal_data": {
-            "dummy": []
-        },
-    },
-    {
-        "prompt_token_ids": [],
-        "multi_modal_data": {
-            "dummy": []
-        },
-    },
-])
-def test_preprocessor_text_no_mm_inputs(model_id, prompt):
-    model_config = ModelConfig(model=model_id)
-    tokenizer = init_tokenizer_from_configs(model_config)
-    input_preprocessor = InputPreprocessor(model_config, tokenizer)
-
-    with pytest.raises(ValueError, match="does not support multimodal inputs"):
-        input_preprocessor.preprocess(prompt)
-
-
-@pytest.mark.parametrize("model_id", [
-    "facebook/chameleon-7b",
-])
-@pytest.mark.parametrize("prompt", [
-    "",
-    {
-        "prompt_token_ids": []
-    },
-])
-def test_preprocessor_always_mm_code_path(model_id, prompt):
-    model_config = ModelConfig(model=model_id)
-    tokenizer = init_tokenizer_from_configs(model_config)
-    input_preprocessor = InputPreprocessor(model_config, tokenizer)
-
-    # HF processor adds sep token
-    sep_token_id = tokenizer.vocab[tokenizer.sep_token]
-
-    processed_inputs = input_preprocessor.preprocess(prompt)
-    assert sep_token_id in processed_inputs["prompt_token_ids"]
